web/backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
composio_core==0.5.49
cryptography==44.0.0
exceptiongroup==1.2.2
fastapi==0.115.5
frozenlist==1.5.0
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
paramiko==3.5.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyNaCl==1.5.0
pyperclip==1.9.0
Pysher==1.0.8
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
semver==3.0.2
sentry-sdk==2.19.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websocket-client==1.8.0
websockets==14.1
yarl==1.18.0
zipp==3.21.0
